Refactor wasm_wrappers: rename some folders; move code to different files; tests now use TypeScript
