+++ .github/workflows/wasm.yml
@@ -35,14 +35,19 @@ jobs:
         uses: actions/setup-node@v3
         with:
           node-version: ${{ matrix.node-version }}
+      - name: Install TypeScript
+        run: npm install typescript
       - name: Install wasm-pack
         run: cargo install wasm-pack
       - name: Build the wasm module
         working-directory: ./wasm-wrappers
         run: wasm-pack build --target nodejs
+      - name: Compile the tests
+        working-directory: ./wasm-wrappers
+        run: tsc --project js-bindings-test/tsconfig.json
       - name: Run the tests
         working-directory: ./wasm-wrappers
-        run: node js-bindings/node-entry.js
+        run: node --enable-source-maps js-bindings-test/node-entry.js
 
   wasm_artifacts:
     runs-on: ubuntu-latest
