+++ do_checks.sh
@@ -5,6 +5,8 @@ set -e
 SCRIPT_DIR=$(cd "$(dirname "$0")" && pwd)
 PYTHON=$(which python || which python3)
 
+cd "$SCRIPT_DIR"
+
 cargo fmt --check -- --config newline_style=Unix
 
 # Install cargo deny first with: cargo install cargo-deny.
@@ -36,7 +38,7 @@ cargo clippy --all-features --workspace --lib --bins --examples -- \
     -D clippy::string_slice
 
 # Install requirements with: pip install -r ./build-tools/codecheck/requirements.txt
-"$PYTHON" "$SCRIPT_DIR/build-tools/codecheck/codecheck.py"
+"$PYTHON" "build-tools/codecheck/codecheck.py"
 
 # Ensure that wasm documentation is up-to-date
 cargo run -p wasm-doc-gen -- -o wasm-wrappers/WASM-API.md --check
