+++ wasm-wrappers/js-bindings-test/node-entry.js
@@ -1,4 +1,4 @@
-import { run_test } from "./wasm_test.js";
+import { run_test } from "./test.js";
 
 async function node_run() {
   await run_test();
