+++ wasm-wrappers/js-bindings-test/package.json
