+++ wasm-wrappers/js-bindings-test/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "target": "es2022",
+    "module": "es2022",
+    "sourceMap": true
+  }
+}
